requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
flask-smorest==0.44.0
Authlib==1.2.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
requests==2.26.0
PyJWT==2.1.0
Flask-Cors==4.0.0
